pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 130

[tool.isort]
profile = "black"
line_length = 130

[tool.ruff]
line-length = 130
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "RUF", "C901"]
ignore = ["E501"]  # Line length handled by black

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
files = ["src"]

[[tool.mypy.overrides]]
module = [
    "tree_sitter",
    "tree_sitter_python",
    "tree_sitter_javascript",
    "tree_sitter_typescript",
    "tree_sitter_typescript.tsx",
    "tree_sitter_go",
    "tree_sitter_rust",
    "tree_sitter_java",
    "tree_sitter_c",
    "tree_sitter_cpp",
    "tree_sitter_ruby",
    "tree_sitter_c_sharp",
    "mistune",
    "pysbd",
    "ruamel.yaml",
    "lxml",
    "lxml.html",
    "tiktoken",
    "sentence_transformers",
    "spacy",
    "spacy.cli",
    "spacy.cli.download",
    "spacy.language",
]
ignore_missing_imports = true

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_files = ["src/treemapper/version.py:__version__"]

[project]
name = "treemapper"
dynamic = ["version"]  # Version is still managed in version.py
authors = [
    { name = "Nikolay Eremeev", email = "nikolay.eremeev@outlook.com" },
]
description = "Export codebase structure and contents for AI/LLM context"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
keywords = ["code-analysis", "directory-tree", "yaml", "json", "llm", "ai", "codebase", "context", "chatgpt", "claude", "code-context", "export", "tree", "gpt-context", "llm-context", "code-to-prompt", "claude-context"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",

    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "pathspec>=0.11,<2.0",
    "pyyaml>=6.0.2,<8.0",
    "tiktoken>=0.7,<1.0",
    "mistune>=3.0,<4.0",
    "ruamel.yaml>=0.18,<1.0",
    "lxml>=5.0,<7.0",
]

[project.urls]
Homepage = "https://github.com/nikolay-e/treemapper"
Repository = "https://github.com/nikolay-e/treemapper"
"Bug Tracker" = "https://github.com/nikolay-e/treemapper/issues"

[project.scripts]
treemapper = "treemapper.treemapper:main"

[project.optional-dependencies]
tree-sitter = [
    "tree-sitter>=0.21,<1.0",
    "tree-sitter-python>=0.21,<1.0",
    "tree-sitter-javascript>=0.21,<1.0",
    "tree-sitter-typescript>=0.21,<1.0",
    "tree-sitter-go>=0.21,<1.0",
    "tree-sitter-rust>=0.21,<1.0",
    "tree-sitter-java>=0.21,<1.0",
    "tree-sitter-c>=0.21,<1.0",
    "tree-sitter-cpp>=0.22,<1.0",
    "tree-sitter-ruby>=0.21,<1.0",
    "tree-sitter-c-sharp>=0.21,<1.0",
]
full = [
    "treemapper[tree-sitter]",
    "pysbd>=0.3,<1.0",
]
embeddings = [
    # Semantic code embeddings for diffctx
    "sentence-transformers>=3.0,<6.0",
]
nlp = [
    # Natural language processing for document tokenization
    "spacy>=3.7,<4.0",
]
dev = [
    "treemapper[tree-sitter]",
    "pytest>=7.0,<10.0",
    "pytest-cov>=3.0,<8.0",
    "pytest-xdist>=3.0,<4.0",
    "hypothesis>=6.0,<7.0",
    "mutmut>=2.0,<4.0",
    "coverage>=7.0,<8.0",
    # Quality checks
    "ruff>=0.4,<1.0",
    "black>=23.0.0,<27.0",
    "isort>=5.12,<8.0",
    "mypy>=1.0,<2.0",
    "radon>=6.0,<7.0",
    "import-linter>=2.0,<3.0",
    "pre-commit>=3.0,<5.0",
    "autoflake>=2.0,<3.0",
    # Type stubs
    "types-PyYAML>=6.0,<7.0",
    # Build and release
    "build>=0.10,<2.0",
    "twine>=4.0,<7.0",
    "pyinstaller>=5.0,<7.0",
    # Security fixes
    "urllib3>=2.6.0",
]

[tool.setuptools.dynamic]
version = { attr = "treemapper.version.__version__" }

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
norecursedirs = ["mutants", ".git", ".mypy_cache", ".pytest_cache", "__pycache__", ".hypothesis"]
addopts = "-n auto --dist loadfile"
